requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
pymysql==1.1.1
cryptography==44.0.2
pyjwt==2.10.1
werkzeug==3.1.3
gunicorn==23.0.0
